requirements.txt
flask==3.0.0
nltk==3.8.1
spacy==3.7.2
tensorflow==2.15.0
torch==2.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.24.3
gensim==4.3.2
PyPDF2==3.0.1
python-docx==1.1.0
wordcloud==1.9.2
werkzeug==3.0.1

